% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{assess_expt_performance_crit}
\alias{assess_expt_performance_crit}
\title{Assesses task performance criteria of the experimental session}
\usage{
assess_expt_performance_crit(data)
}
\arguments{
\item{data}{A tibble containing block data from experiment session}
}
\description{
Returns a tibble with columns:
subjectIx
criterion
mean_perf_overall
failed_overall_crit
max_n_consec_fails
failed_blockwise_crit
}
