% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_dir}
\alias{check_dir}
\title{Check whether directories exists, and create recursively if they don't}
\usage{
check_dir(all_dirs)
}
\arguments{
\item{all_dirs}{Character vector of directory paths to check}
}
\description{
Check whether directories exists, and create recursively if they don't
}
