% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{compute_ssrt}
\alias{compute_ssrt}
\title{Compute stop-signal reactione time}
\usage{
compute_ssrt(tibb, for_each_ssd = FALSE)
}
\description{
Stop-signal reaction time (SSRT) can be computed according to one of the
following methods:
- integration method of the independent race model (Logan & Cowan, Psych Rev,
1984)
- mean method (Logan & Cowan, Psych Rev, 1984)
- modified integration method ()
}
\details{
Integration method ----------------------------------------------------------
- Compute probability of responding given a stop-signal across delays (staircase) or for each delay separately
(no staircase)
- Estimate finishing time: integrate the no-signal response time distribution up to the point corresponding to the probability of responding given a stop-signal
- Subtract mean delay (staircase) or delay (no staircase)

Mean method -----------------------------------------------------------------
- Compute probability of responding given a stop-signal across delays (staircase) or for each delay separately
(no staircase)
- Estimate finishing time: mean no-signal response time

Modified integration method -------------------------------------------------
- Compute probability of responding
}
