% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{derive_bayesian_logistic_regression_preds}
\alias{derive_bayesian_logistic_regression_preds}
\title{Derive Bayesian logistic regression predictionsgiven best-fitting parameters}
\usage{
derive_bayesian_logistic_regression_preds(tibb)
}
\arguments{
\item{tibb}{Tibble containing columns subject identifier (subjectIx), best-fitting parameters (mean_beta0, mean_beta), and Bayes factor (B)}
}
\description{
Derive Bayesian logistic regression predictionsgiven best-fitting parameters
}
