% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{fancy_scientific}
\alias{fancy_scientific}
\title{Convert double to scientific notation-formatted string for annotation purposes}
\usage{
fancy_scientific(dbl)
}
\arguments{
\item{dbl}{double}
}
\description{
Convert double to scientific notation-formatted string for annotation purposes
}
