% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{get_bf_colorbar_vars}
\alias{get_bf_colorbar_vars}
\title{Get colorbar variables for visualizaing Bayes factor data}
\usage{
get_bf_colorbar_vars()
}
\description{
Get colorbar variables for visualizaing Bayes factor data
}
