% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{get_bf_settings}
\alias{get_bf_settings}
\title{Get parameters for Bayesian hypothesis testing}
\usage{
get_bf_settings(param)
}
\arguments{
\item{param}{parameter to get settings of}
}
\description{
These include the width of the priod (rscale), Bayes factor category labels (labels) and cutoffs (breaks).
}
