% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{get_col_types}
\alias{get_col_types}
\title{Get column types for extracting data from csv files}
\usage{
get_col_types(file_type)
}
\arguments{
\item{file_type}{Type of file to be read}
}
\description{
File type can be any of the following:
}
\details{
expt_block_data - tidied block-level data from experimental session
expt_trial_resp_data - tidied trial-level response data from experimental session
expt_trial_rt_data - tidied trial-level resonse time data from experimental session
log_sess_cols - source data, session-related columns
log_block_cols - source data, block-related columns
log_trial_cols - source data, trial-related columns
log_trial_cols_sub00 - source data, trial-related columns for subject 00
prac_block_data - tidied block-level data from practice session
}
