% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_figshare_specs}
\alias{get_figshare_specs}
\title{Return project-specific variable values for uploading to figShare}
\usage{
get_figshare_specs(spec, article_id)
}
\arguments{
\item{spec}{variable}

\item{article_id}{figShare ID number of the article (optional)}
}
\description{
Return project-specific variable values for uploading to figShare
}
