% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{identify_outliers_analysis_level}
\alias{identify_outliers_analysis_level}
\title{Identify analysis-level outliers}
\usage{
identify_outliers_analysis_level(df, analysis_id, trial_alt_levels)
}
\arguments{
\item{df}{data frame with summary stats for each subject}

\item{analysis_id}{analysis identifier (e.g. 'a5'), corresponds to analysis ID in preregistration document}

\item{trial_alt_levels}{trial_alt levels to include}
}
\description{
Identify analysis-level outliers
}
