% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{identify_outlying_values}
\alias{identify_outlying_values}
\title{Identify analysis-level outliers}
\usage{
identify_outlying_values(df, group_vars, dep_var, z_threshold = 2.5)
}
\arguments{
\item{df}{data frame with summary stats for each subject}

\item{group_vars}{variables for grouping}

\item{dep_var}{dependent variables}

\item{z_threshold}{outlier criterion z score (defaults to 2.5)}
}
\description{
Identify analysis-level outliers
}
