% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{join_stats_tibbles}
\alias{join_stats_tibbles}
\title{Join tibbles containing descriptive and inferential statistics data}
\usage{
join_stats_tibbles(descriptives, inferentials)
}
\arguments{
\item{descriptives}{tibble containing descriptives statistics}

\item{inferentials}{tibble containing inferential statistics}
}
\description{
Joining occurs via a left-join operation by the subjectIx variable
}
