% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_if_grp}
\alias{plot_if_grp}
\title{Plot inhibition function for group-level data}
\usage{
plot_if_grp(obs, prd)
}
\arguments{
\item{obs}{tibble containing observations}

\item{prd}{tibble containing predicted parameter estimates from Bayesian logistic regression model}
}
\description{
Plot inhibition function for group-level data
}
