% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_if_idv}
\alias{plot_if_idv}
\title{Plot inhibition function for individual-level data}
\usage{
plot_if_idv(obs_data, prd_data, bf_data, colorbar, yvar, bvar)
}
\arguments{
\item{obs_data}{tibble with observed response probabilities}

\item{prd_data}{tibble with predicted response probabilities}

\item{bf_data}{Bayes factor data}

\item{colorbar}{Color bar variables}

\item{yvar}{y axis variable}

\item{bvar}{Bayes factor variable}
}
\description{
Plots the probability of responding given a stop-signal against stop-signal delay, a histogram of the number stop-inhibit and stop-respond trials for each stop-signal delay, and a logistic regression fit to the data.
}
