% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_mcmc_analysis}
\alias{plot_mcmc_analysis}
\title{plot_mcmc_analysis ##########################################################
Plots graphical diagnostics of MCMC chains (for models fit with brms)}
\usage{
plot_mcmc_analysis(mdl)
}
\arguments{
\item{mdl, }{brmsfit object}
}
\description{
plot_mcmc_analysis ##########################################################
Plots graphical diagnostics of MCMC chains (for models fit with brms)
}
