% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_overview}
\alias{plot_overview}
\title{Plots task performance overview, for assessment of performance criteria}
\usage{
plot_overview(data, xvar, xvardesc, yorder, yticklabels, fillvar, fillvardesc,
  labelvar, facetwrapvar, facetgridvar, plottitle)
}
\arguments{
\item{data}{tibble containing data}

\item{xvar}{x axis variable}

\item{xvardesc}{x axis variable}

\item{yorder}{criterion level order}

\item{yticklabels}{criterion level description}

\item{fillvar}{fill variable}

\item{fillvardesc}{fill variable description}

\item{labelvar}{label variable}

\item{facetwrapvar}{facet_wrap variable(s), if any}

\item{facetgridvar}{facet_grid variable(s), if any
@param plottitle plot title, if any}
}
\description{
Plotting can be done for practice and experimental stage.
}
