% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_posterior}
\alias{plot_posterior}
\title{plot_posterior ##########################################################
Plots posterior distribution}
\usage{
plot_posterior(mdl)
}
\arguments{
\item{mdl, }{brmsfit object}
}
\description{
plot_posterior ##########################################################
Plots posterior distribution
}
