% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_srrt_vs_nsrt_grp}
\alias{plot_srrt_vs_nsrt_grp}
\title{Plot stop-respond RT vs. no-signal RT for group-level data}
\usage{
plot_srrt_vs_nsrt_grp(data, sr, ns)
}
\arguments{
\item{data}{tibble in wide-format (each row is a subject) containing RT data}

\item{sr}{column name containing stop-respond response times}

\item{ns}{column name containing no-stop response times}
}
\description{
Plot stop-respond RT vs. no-signal RT for group-level data
}
