% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_srrt_vs_ssd_grp}
\alias{plot_srrt_vs_ssd_grp}
\title{Plot stop-respond RT vs. stop-signal delay for group-level data}
\usage{
plot_srrt_vs_ssd_grp(df)
}
\description{
Plot stop-respond RT vs. stop-signal delay for group-level data
}
