% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{rank_by_bayes_factor}
\alias{rank_by_bayes_factor}
\title{Rank subjects based on the Bayes Factor}
\usage{
rank_by_bayes_factor(tibb, bvar = "B")
}
\arguments{
\item{tibb}{Tibble containing columns named subjectIx and B (Bayes Factors)}
}
\description{
Rank subjects based on the Bayes Factor
}
