% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{read_log_files}
\alias{read_log_files}
\title{Reads logs files}
\usage{
read_log_files(files, data_type)
}
\arguments{
\item{files}{files: list of files}

\item{data_type}{Data type (char): trial_data, block_data, sess_data}
}
\description{
Reads logs files
}
