% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_functions.R
\name{render_notebook}
\alias{render_notebook}
\title{Batch-render analysis notebooks}
\usage{
render_notebook(notebook_file, notebook_dir = "analysis",
  reports_dir = "reports", params_tibble, force = FALSE)
}
\arguments{
\item{notebook_file}{filename of the template notebook to be run}

\item{notebook_dir}{directory where the template notebook resides}

\item{reports_dir}{directory where reports are written}

\item{params_tibble}{tibble of parameter values with which to run the notebooks}

\item{force}{whether or note to rerun a notebook when it exists}
}
\description{
Batch-render analysis notebooks
}
