% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{savage_dickey}
\alias{savage_dickey}
\title{Savage-Dicky density ratio test}
\usage{
savage_dickey(samples, prior = dnorm(0, 1/2^2))
}
\arguments{
\item{samples}{MCMC samples}

\item{prior}{prior}
}
\description{
Savage-Dicky density ratio test
}
