% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{select_log_files}
\alias{select_log_files}
\title{WHAT IT DOES}
\usage{
select_log_files(stage = "experiment", filetype = "triallog")
}
\arguments{
\item{stage}{Study stage: practice or experiment}

\item{filetype}{File type: triallog or blocklog}
}
\description{
Selects performance log files from raw data directory
}
\details{
Selection of performance log files based on stage ("practice", "experiment") and file type ("trial", "block").
}
