% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{show_trial_numbers}
\alias{show_trial_numbers}
\title{Show trial numbers for al trial types and all individuals}
\usage{
show_trial_numbers(df)
}
\arguments{
\item{df}{a tidied data frame containing the performance data}
}
\description{
Show trial numbers for al trial types and all individuals
}
