% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{split_log_in_multi_table}
\alias{split_log_in_multi_table}
\title{Splits log files into a session table and a block/trial table}
\usage{
split_log_in_multi_table(df)
}
\description{
Splits log files into a session table and a block/trial table
}
