% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{test_effect_ssd_on_srrt_grp}
\alias{test_effect_ssd_on_srrt_grp}
\title{Test of stop-respond RT vs. stop-signal delay for group-level data}
\usage{
test_effect_ssd_on_srrt_grp(df)
}
\arguments{
\item{df}{data frame}
}
\description{
Test of stop-respond RT vs. stop-signal delay for group-level data
}
