% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{test_effect_ssd_on_srrt_idv}
\alias{test_effect_ssd_on_srrt_idv}
\title{Test of stop-respond RT vs. stop-signal delay for individual-level data}
\usage{
test_effect_ssd_on_srrt_idv(tibb, rscale = get_bf_settings("rscale"))
}
\arguments{
\item{tibb}{a tibble containing the data for the analysis}
}
\description{
Bayesian ANOVA
}
