% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{test_if_grp}
\alias{test_if_grp}
\title{Test of inhibition function for group-level data}
\usage{
test_if_grp(tibb, stopping_type, derivatives_dir, figures_dir,
  notebook_name)
}
\arguments{
\item{tibb}{tibble containing subject-level data, including Bayesian logistic regression parameter estimates}

\item{par_name}{name of the parameter containing the estimate of the logistic regression slope}
}
\description{
Tests the independent race model's qualitative prediction that the probability of responding given a stop-signal increases as a function of stop-signal delay, in group-level data.
}
