% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{test_if_idv}
\alias{test_if_idv}
\title{Test of inhibition function for individual-level data}
\usage{
test_if_idv(tibb, stopping_type, derivatives_dir, figures_dir,
  notebook_name, modeling_approach = "brms", rscale = 0.5)
}
\arguments{
\item{tibb, }{tibble containing data on subjectIx trial_alt t_d and r_bi}

\item{figures_dir, }{path to the directory for saving figures}

\item{notebook_name, }{name of the notebook}

\item{modeling_approach, }{whether using kruscke methods or brms to fit models}

\item{rscale, }{width of the prior (in standardized effect size units)}
}
\description{
Tests the independent race model's qualitative prediction that the probability of responding given a stop-signal increases as a function of stop-signal delay, in individual-level data.
}
