% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{test_srrt_vs_nsrt_grp}
\alias{test_srrt_vs_nsrt_grp}
\title{Test of stop-respond RT vs. no-signal RT for group-level data}
\usage{
test_srrt_vs_nsrt_grp(data, sr, ns)
}
\arguments{
\item{data}{a tibble}

\item{sr}{column name of stop-respond trial response time data}

\item{ns}{column name of no-stop trial response time data}
}
\description{
Paired t-test
}
