% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{test_srrt_vs_nsrt_idv}
\alias{test_srrt_vs_nsrt_idv}
\title{Test of stop-respond RT vs. no-signal RT for individual-level data}
\usage{
test_srrt_vs_nsrt_idv(tibb, trial_alt_levels,
  rscale = get_bf_settings("rscale"))
}
\arguments{
\item{tibb}{tibble with trial-level data}

\item{trial_alt_levels}{2-element character vector containing the levels to contrast}
}
\description{
Tests the independent race model's qualitative prediction that stop-respond RT should be shorter than no-signal, in individual-level data.
}
