% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{theme_irmass}
\alias{theme_irmass}
\title{Get default ggplot theme settings to be used in this project}
\usage{
theme_irmass()
}
\description{
Get default ggplot theme settings to be used in this project
}
