% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{tidy_block_data}
\alias{tidy_block_data}
\title{Tidy block data}
\usage{
tidy_block_data(df, stage = "expt")
}
\description{
Tidying involves multiple steps, including aggregating no-signal performance trial performance across go-signal levels, renaming variables, and assessing performance criteria on the basis of aggregate measures.
}
