% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{upload_data_to_figshare}
\alias{upload_data_to_figshare}
\title{Whether or not to upload data to figShare}
\usage{
upload_data_to_figshare()
}
\description{
When this function returns TRUE, then data and figures are uploaded to figShare, DOIs are created, data is made public.
}
