% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{verify_ix_cols}
\alias{verify_ix_cols}
\title{Verifies index columns}
\usage{
verify_ix_cols(df)
}
\description{
Identifies index columns, replaces NaNs with NAs, and converts objects to integer
}
