% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{verify_output_dirs}
\alias{verify_output_dirs}
\title{Checks if directories for notebook output exists, and if not creates them}
\usage{
verify_output_dirs(base_dirs, notebook_name)
}
\arguments{
\item{base_dirs}{A list of directories in which subdirectories should be created, if they don't exist}

\item{notebook_name}{Name of subdir}
}
\value{
created_any Whether or not any output dir was created
}
\description{
Checks if directories for notebook output exists, and if not creates them
}
