% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{write_fig_cap}
\alias{write_fig_cap}
\title{Write figure caption to disk}
\usage{
write_fig_cap(fig_title, fig_description, fig_dir, notebook_name,
  stopping_type)
}
\description{
Write figure caption to disk
}
