% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{test_if_idv}
\alias{test_if_idv}
\title{Test of inhibition function for individual-level data}
\usage{
test_if_idv(tibb)
}
\arguments{
\item{data}{a data frame containing the data; y variable should be coded as integers}

\item{y_name}{e.g. 'r_bi'}

\item{x_name}{e.g. 't_d'}

\item{num_saved_steps}{number of saved steps}

\item{thin_steps}{how many steps for thinning}

\item{file_name_root}{base file name for output}

\item{graph_file_type}{graphic output format (defaults to 'eps')}
}
\description{
Tests the independent race model's qualitative prediction that the probability of responding given a stop-signal increases as a function of stop-signal delay, in individual-level data.
}
