% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{tidy_trial_data}
\alias{tidy_trial_data}
\title{Tidy trial data}
\usage{
tidy_trial_data(df)
}
\description{
Tidying involves the following steps
- renaming and computing variables in line with preregistration
}
